refactor(hub): roadmap/changelog: extract hook, colors; add rq provider
ref: #MANAGER-19402
