+++ packages/manager/apps/hub/src/__mocks__/ReactQueryWrapper.tsx
@@ -1,0 +1,14 @@
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { PropsWithChildren } from 'react';
+
+export function ReactQueryWrapper({
+  children,
+  client,
+}: PropsWithChildren<{ client?: QueryClient }>) {
+  const queryClient = client ?? new QueryClient();
+  return (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
+}
+
+export default ReactQueryWrapper;
