+++ packages/manager/apps/hub/src/components/roadmap-changelog-datagrid/RoadmapChangelogDatagrids.tsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { Datagrid, DatagridColumn } from '@ovh-ux/manager-react-components';
+import { Datagrid } from '@ovh-ux/manager-react-components';
 import {
   OsdsTabBarItem,
   OsdsTabPanel,
@@ -7,7 +7,6 @@ import {
   OsdsLink,
   OsdsIcon,
   OsdsTabBar,
-  OsdsSkeleton,
 } from '@ovhcloud/ods-components/react';
 import {
   ODS_TABS_SIZE,
@@ -25,16 +24,9 @@ import {
 import { useTranslation } from 'react-i18next';
 import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
 import { EXTERNAL_LINKS, ROADMAP_CHANGELOG_PAGES } from '@/changelog.constants';
-import {
-  RoadmapChangelogItemProductCell,
-  RoadmapChangelogItemTitleCell,
-  RoadmapChangelogItemDescriptionCell,
-  RoadmapChangelogItemReleaseDateCell,
-  RoadmapChangelogItemStatusCell,
-} from './cells';
 import { useRoadmapChangelog } from '@/data/hooks/roadmapChangelog/useRoadmapChangelog';
 import styles from './style.module.scss';
-import { RoadmapChangelogItem } from '@/types/roadmapchangelog.type';
+import useRoadmapChangelogData from '@/hooks/roadmapChangelog/useRoadmapChangelogData';
 
 const RoadmapChangelogDatagrids = () => {
   const {
@@ -46,72 +38,9 @@ const RoadmapChangelogDatagrids = () => {
   const { environment } = useContext(ShellContext);
   const isRegionUS = environment.getRegion() === 'US';
 
-  const emptyRoadmapChangelogItem: RoadmapChangelogItem = {
-    title: '',
-    description: '',
-    product: '',
-    releaseDate: '',
-    status: '',
-    changelog: '',
-  };
-  const emptyRoadmapChangelogItems: RoadmapChangelogItem[] = [
-    emptyRoadmapChangelogItem,
-    emptyRoadmapChangelogItem,
-    emptyRoadmapChangelogItem,
-  ];
-
-  const columns: DatagridColumn<RoadmapChangelogItem>[] = [
-    {
-      id: 'product',
-      label: t('datagrid_table_head_product'),
-      cell: (item: RoadmapChangelogItem) =>
-        isLoadingItems ? (
-          <OsdsSkeleton />
-        ) : (
-          <RoadmapChangelogItemProductCell item={item} />
-        ),
-    },
-    {
-      id: 'changelog',
-      label: t('datagrid_table_head_changelog'),
-      cell: (item: RoadmapChangelogItem) =>
-        isLoadingItems ? (
-          <OsdsSkeleton />
-        ) : (
-          <RoadmapChangelogItemTitleCell item={item} />
-        ),
-    },
-    {
-      id: 'description',
-      label: t('datagrid_table_head_description'),
-      cell: (item: RoadmapChangelogItem) =>
-        isLoadingItems ? (
-          <OsdsSkeleton />
-        ) : (
-          <RoadmapChangelogItemDescriptionCell item={item} />
-        ),
-    },
-    {
-      id: 'realease_date',
-      label: t('datagrid_table_head_release_date'),
-      cell: (item: RoadmapChangelogItem) =>
-        isLoadingItems ? (
-          <OsdsSkeleton />
-        ) : (
-          <RoadmapChangelogItemReleaseDateCell item={item} />
-        ),
-    },
-    {
-      id: 'status',
-      label: t('datagrid_table_head_status'),
-      cell: (item: RoadmapChangelogItem) =>
-        isLoadingItems ? (
-          <OsdsSkeleton />
-        ) : (
-          <RoadmapChangelogItemStatusCell item={item} />
-        ),
-    },
-  ];
+  const { emptyRoadmapChangelogItems, columns } = useRoadmapChangelogData(
+    isLoadingItems,
+  );
 
   return (
     <div className="max-w-[100%] overflow-auto">
