+++ packages/manager/apps/hub/src/data/hooks/roadmapChangelog/useRoadmapChangelog.spec.tsx
@@ -1,16 +1,14 @@
 import { PropsWithChildren } from 'react';
 import { renderHook, waitFor } from '@testing-library/react';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { describe, it, vi } from 'vitest';
 import { aapi as Api } from '@ovh-ux/manager-core-api';
 import { useRoadmapChangelog } from '@/data/hooks/roadmapChangelog/useRoadmapChangelog';
 import { ApiEnvelope } from '@/types/apiEnvelope.type';
 import { RoadmapChangelogResponse } from '@/types/roadmapchangelog.type';
-
-const queryClient = new QueryClient();
+import { ReactQueryWrapper } from '@/__mocks__/ReactQueryWrapper';
 
 const wrapper = ({ children }: PropsWithChildren) => (
-  <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  <ReactQueryWrapper>{children}</ReactQueryWrapper>
 );
 
 describe('useRoadmapChangelog', () => {
